+++ domain/src/main/java/com/agileburo/anytype/domain/block/model/Command.kt
@@ -178,11 +178,13 @@ sealed class Command {
      * Command for splitting one block into two blocks
      * @property context context id
      * @property target id of the target block, which we need to split
+     * @property style target block text style
      * @property index index or cursor position
      */
     data class Split(
         val context: Id,
         val target: Id,
+        val style: Block.Content.Text.Style,
         val index: Int
     )